# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Actor:
    properties:
      client_id:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  response.Meta:
    properties:
      actor:
        $ref: '#/definitions/response.Actor'
      request_id:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/response.ResponseError'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ResponseError:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  reviews.CreateReviewRequest:
    properties:
      content:
        type: string
      entity_id:
        type: string
      entity_type:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - content
    - entity_id
    - entity_type
    - rating
    - title
    type: object
  reviews.ReviewResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      entity_id:
        type: string
      entity_type:
        type: string
      id:
        type: string
      rating:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  reviews.ReviewsListResponse:
    properties:
      average_rating:
        type: number
      limit:
        type: integer
      page:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/reviews.ReviewResponse'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  reviews.UpdateReviewRequest:
    properties:
      content:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - content
    - rating
    - title
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@gogin.com
    name: API Support
  description: |-
    A comprehensive REST API built with Go and Gin framework
    This API provides user authentication, profile management, and admin operations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gogin API
  version: "1.0"
paths:
...
  /reviews:
    get:
      parameters:
      - description: Entity type
        in: query
        name: entity_type
        required: true
        type: string
      - description: Entity ID
        in: query
        name: entity_id
        required: true
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/reviews.ReviewsListResponse'
              type: object
      summary: List Reviews
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Review details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reviews.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/reviews.ReviewResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create Review
      tags:
      - Reviews
  /reviews/{id}:
    delete:
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete Review
      tags:
      - Reviews
    get:
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/reviews.ReviewResponse'
              type: object
      summary: Get Review
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Review update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reviews.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/reviews.ReviewResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update Review
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...